+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/3.2.19...HEAD)
+## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/3.2.20...HEAD)
+
+## [3.2.20](https://github.com/alexdlaird/amazon-orders/compare/3.2.19...3.2.20) - TBD
+
+### Added
+
+- Support for Amazon's new login auth flow. Removed `SIGN_IN_REDIRECT_URL` as part of this, since it is no longer necessary.
+- Improvements to error handling and logging when persisted session expires.
+- Stability improvements.
 
 ## [3.2.19](https://github.com/alexdlaird/amazon-orders/compare/3.2.18...3.2.19) - 2025-04-21